Renamed "rapcom" to "rcli".
