+++ .travis.yml
@@ -14,7 +14,7 @@ deploy:
   provider: pypi
   user: dangle
   password:
-    secure: N+iuk3Vn4zKWrDeK5ln7BKrs9EbSfoAqGc+zbeaEHcWn+wNXmdp9poNdy14pIKzrPXBfZiEy7MLDBzmGwvi9BkEivEV21tS8FJ45p/Rm/1HhNc1LygfiIeaO1NivYluzIOsczutKXOoDqvGtqc2UYPOuvd7ybosLpjMPGG6bLd8gAHMCaBuI5nl17alB8igdrhXBMbpBiir/7tKlaYfZvO43Ix3UULExkdQHdZBLpSFdgG11Fv9cq8jyrLZtdmXzae0Yu0og+OPl1AGUtR6I4piBShYzN29s6AXQv+FcdORGtfIKzbPGjDlYed5URp0Q00m7mojeqnCawqzWxMhl2wKGvkI+f4ybRhjHTAb70DzJbvjOiSC6lwbvOFs3mp65iR5lEXtJ4+5dM70TSrk8hGGRhHKYf/ZuNFgr8EdaHna8SVTex67EB7dbb0b4oyzeDVCRFYaSCxEh7h3xnlyv3BwJ4TdZ6wshOn2KNf/Ritr3LOXjwJ9VCsahJ3EIoV7YEzsLcLCgpt4RVToEuHdGmEmSSFwxYfVfD0CPvHwHrTPunq/gtfFwtKum4gpgvgWcUQYH+LoHXjjco4bVZL3p9qsbee22tfAF2Ke9ldvO3MMKuJE4/3LA2/iWLH1Xj+rsDS6FqGQ9GLmejvXaZW47xUPyxVR1Sv49pVgnQJJIyWo=
+    secure: PgVFsaNvHX2c8rm0JHvypFYC4cnIvozr8OdGSl/AUdOAswffYBGgUatxJ4O7JbKvC1Ig0nIj5C6XAgxPuw9qsJ+Xkfdo1htVF6AEVPD/1n4rHstO3bToZV0STOJYZbKpPuugr2Y6lXlMDUKNjfsPq0A6KLaksXSvkW3/LZ81zncW5b+jMrfndZRuA64xN59qEAk2vFsvTYaVQOXJ9JgSTJ3OdTmcbTm1x+jzhDXqEL64RHDNLt+Ca5JSNagfSYyNx85J/3l36t76E4GH2pD3NEIfaIO2EQmqk1CPknhdUaGZUpVBP+6vitmrWShzVBSNgBbdTwonyYpqdrqp/YcIDSzD+QuLbrOLOmXRCDVTmmBwXp5+0wcDxolCXKFltrn5PdINBas70gAkiv3IKGMjggbD5vgZhp3hPMHSwalPtcAIFLIFFI2mXXVSyxJqtBz2pmSNhwdaFyeUU9hSC4/02Y9nqVePoxfVMAeybPnMckLFLW4EDgp1tNoepugMt8RSEQTMRcrTvkjZ9FtPNCHzAC909pLHqBHvK4CVx9+uxQQIR8peeg/bSODcvMYZ2frFkLiv/xBa5xL3Z57juKdEzRxycbrUzVBQML/KOUwL3WPSb8xk3NRD+69WLyzudFWuiShgtBcSZ2GWC/c1bUQKqH0bY8rUhOoqkh7XpHVFiK0=
   on:
     tags: true
-    repo: containenv/rapcom
+    repo: containenv/rcli
