+++ examples/simple_example/setup.py
@@ -4,7 +4,6 @@ setup(
     name='simple_example',
     version='1.0.0',
     packages=['simple_example'],
-    install_requires=['rapcom'],
-    setup_requires=['rapcom'],
+    setup_requires=['rcli'],
     autodetect_commands=True
 )
