--- examples/validation_example/validation_example/__init__.py
@@ -1,18 +1,0 @@
-def hello():
-    """usage: say hello"""
-    print('Hello!')
-
-
-def hiya(name):
-    """usage: say hiya <name>"""
-    print('Hiya, {name}!'.format(name=name))
-
-
-def validate_name(**cli_args):
-    assert '--name' in cli_args
-    assert str(cli_args['--name'])
-    assert len(cli_args['--name'])
-    return cli_args
-
-
-hiya.validate = validate_name
