+++ rcli/autodetect.py
@@ -41,9 +41,9 @@ def setup_keyword(dist, _, value):
         dist.entry_points = {}
     for command, subcommands in six.iteritems(_get_commands(dist)):
         dist.entry_points.setdefault('console_scripts', []).append(
-            '{command} = rapcom.dispatcher:main'.format(command=command)
+            '{command} = rcli.dispatcher:main'.format(command=command)
         )
-        dist.entry_points.setdefault('rapcom', []).extend(subcommands)
+        dist.entry_points.setdefault('rcli', []).extend(subcommands)
 
 
 def _get_commands(dist):
