+++ rcli/dispatcher.py
@@ -186,7 +186,7 @@ def _get_subcommands():
     """Return all subcommands for the current command."""
     regex = re.compile(r'{}:(?P<name>[^:]+)$'.format(_COMMAND))
     subcommands = {}
-    for ep in pkg_resources.iter_entry_points(group='rapcom'):
+    for ep in pkg_resources.iter_entry_points(group='rcli'):
         try:
             if ep.name == _COMMAND:
                 subcommands[None] = ep.load()
