+++ tox.ini
@@ -17,18 +17,18 @@ deps =
     lint: flake8
     lint: pylint
 commands =
-    test: py.test --cov=rapcom --basetemp={envtmpdir} {posargs}
+    test: py.test --cov=rcli --basetemp={envtmpdir} {posargs}
     test: - coveralls
-    lint: pep8 setup.py rapcom tests
-    lint: pep257 setup.py rapcom tests
-    lint: flake8 setup.py rapcom tests
+    lint: pep8 setup.py rcli tests
+    lint: pep257 setup.py rcli tests
+    lint: flake8 setup.py rcli tests
     lint: pylint \
     lint:     --reports n \
     lint:     -d invalid-name \
     lint:     -d no-member \
     lint:     -d locally-disabled \
     lint:     -d redefined-outer-name \
-    lint:     setup.py rapcom tests
+    lint:     setup.py rcli tests
 
 [flake8]
 ignore = D203
